backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-rest-knox==4.1.0
drf-yasg[validation]
psycopg2-binary==2.9.1
